+++ NEWS.md
@@ -1,3 +1,5 @@
+* C type `long` is used for array dimensions, number of elements and indices.
+
 * Error messages are correctly printed in Julia interface.
 
 * Anonymous semaphores are used to signal new images.